refactor: move history comparison logic into app submodule
